PY-10985/IDEA-104596 extract ConsoleHistoryModel interface
